Show temperature on display
The temperature is now constantly updated on display with each reading.

+++ pico/graphics.h
@@ -5,6 +5,6 @@
 #include <stdio.h>
 
 void graphicsInit();
-void graphicsSetTemperature(uint8_t temperature);
+void graphicsSetTemperature(float temperature);
 
 #endif
